Consolidate Project lifecycle module (#224)
Merge lib/lifecycle-display.ts into lib/portfolio.ts so the Project
type, the operational→public-stage rollup, the labels/titles/colors,
and the type guards live behind one import. Per the deletion test in
the issue: locality improves, no duplication appears elsewhere.

- lib/portfolio.ts gains the Lifecycle Display section (labels,
  PUBLIC_STAGE_CHIP, PUBLIC_STAGE_TITLE, OPERATIONAL_LABEL,
  OPERATIONAL_TITLE, STAGE_OPERATIONAL_ROLLUP, STATUSES,
  isProjectStatus, publicStageFromStatus, stageBreakdown).
- All 7 consumers now import lifecycle helpers from lib/portfolio.
- lib/lifecycle-display.ts deleted.
- CLAUDE.md and .impeccable.md updated to point at the consolidated
  module.

No behavior change. ADR 0001 rules unchanged. The single
implementation of the rollup is computePublicStage; publicStageFromStatus
is the string-tolerant wrapper that delegates to it.

Closes #224.
