+++ .impeccable.md
@@ -237,7 +237,7 @@ would dilute the signal.
 ## Lifecycle Status Treatment
 
 Defined in [ADR 0001](docs/adr/0001-product-lifecycle-taxonomy.md);
-implemented via [`lib/lifecycle-display.ts`](lib/lifecycle-display.ts).
+implemented in [`lib/portfolio.ts`](lib/portfolio.ts) (Lifecycle Display section).
 Two layers, two visual treatments.
 
 **Two-chip cluster on portfolio cards.** Stacked, top-right of the
